Contributing and templates
This is the first stab at a contribution guide, and templates for issues and pull requests.  Its in need of input, please feel free to pull down, amend /  adjust as you see fit.

It would be great to get these done so that we can start stadardising the use or git nad github across the organisation.